CameraServiceProxy: Notify audio service of camera facing
When a camera is opened, notify the audio service which way the camera
faces.  This can be used by the audio system to select microphone
tuning for the CAMCORDER option that's dependent on camera direction.

External facing is treated the same as front-facing.

Bug: 118022272
Test: Manual
Change-Id: I0a89d82a015f371580d335b232013ece80bfa4d3
(cherry picked from commit 790d6fa687eeb77785c2ec08051e94d5647f1322)