Enable @MissingMethodFlags annotation in IMMS
This is a preparation to split
com.android.internal.view.InputMethodClient.

When InputConnectionInspector.MissingMethodFlags annotation was
originally introduced [1], somehow we could not use that annotation in
InputMethodManagerService (IMMS) probably because of some build system
or compiler limitation since the attribute belongs to frameworks.jar
while IMMS belongs to services.jar.

Now it seems that such a limitation no longer exists, there is no
reason to not enable @MissingMethodFlags annotation in IMMS.

@MissingMethodFlags is @Retention(SOURCE) anyway. Hence there should
be no behavior difference at run time.

 [1]: I3c58fadd924fad72cb984f0c23d3099fd0295c64
      2d45ccf90d0745ec7d7e7e441a5809e6aab848c1

Bug: 118040692
Test: compile
Change-Id: I2ecf7a452b088df48f847227faac36062c3b9bc0