Enable @StartInputReason annotation in IMMS
This is a preparation to split
com.android.internal.view.InputMethodClient.

When InputMethodClient.StartInputReason annotation was originally
introduced [1], somehow we could not use that annotation in
InputMethodManagerService (IMMS) probably because of some build system
or compiler limitation since the attribute belongs to frameworks.jar
while IMMS belongs to services.jar.

Now it seems that such a limitation no longer exists, there is no
reason to not enable @StartInputReason annotation in IMMS.

@StartInputReason is @Retention(SOURCE) anyway. Hence there should
be no behavior difference at run time.

 [1]: I74e83c8ca9d1d53e31e9c7b5bda1dec6274e59c8
      ba4de694cb22b8a6790637c75973f4331d2d9a98

Bug: 118040692
Test: compile
Change-Id: I6a7fef9d72bd6aafe4111abc06f8887e9a761236