Enable @UnbindReason annotation in IMMS
This is a preparation to split
com.android.internal.view.InputMethodClient.

When InputMethodClient.UnbindReason annotation was originally
introduced [1], somehow we could not use that annotation in
InputMethodManagerService (IMMS) probably because of some build system
or compiler limitation since the attribute belongs to frameworks.jar
while IMMS belongs to services.jar.

Now it seems that such a limitation no longer exists, there is no
reason to not enable @UnbindReason annotation in IMMS.

@UnbindReason is @Retention(SOURCE) anyway. Hence there should
be no behavior difference at run time.

 [1]: I74e83c8ca9d1d53e31e9c7b5bda1dec6274e59c8
      ba4de694cb22b8a6790637c75973f4331d2d9a98

Bug: 118040692
Test: compile
Change-Id: Ie5af7249da80a2217a4954e898d956ba29fea449