Use raw MotionEvent coordinates when tracking velocity
Unfortunately, we have to roll back the commit "use raw coordinates in
velocitytracker" ag/3545962 because it broke some assumptions in the
view hierarchy. That is being done in aosp/800996.

As a result, we have to start working around the revert by manually
adjusting the coordinates. This is effectively a revert of ag/4071758
that has been rebased on ag/4891556.

Test: manual interaction with notification panel (pulling down,
flinging at various points)
Bug: 117921784
Bug: 117475766

Change-Id: I24c5815967381cd4be983ef7e8456fa7d42c53c8