Add ACL broadcast for bt devices
When we receive profile state change, the device may still be physically
connected(e.g. BluetoothDevice.isConnected() return true). In this case
UI still show a device is connected.

Add ACL broadcast listener which will be triggered when device is
physically disconnected. In this callback we can refresh UI to make it
show correct information.

Bug: 79947085
Test: RunSettingsLibRoboTests
Change-Id: I8370cc568fcb0d5995f164e7e62391a9525cbe19