KeyguardDisableHandler: make properly user aware
Also fixes an issue where the disable handler was not
properly updated after adding a secure lockscreen.

Also fixes an issue where the disable handler was not
properly updated after the device policy changes for
users other than USER_SYSTEM.

Also prevents adding new privileged usages of the API.

Also removes a workaround that prevented Keyguard from
re-engaging if it timed out while the it was disabled.
The workaround is no longer necessary because the in-call
UI is now using the SHOW_WHEN_LOCKED API instead of disabling
the keyguard.

Change-Id: Ib2644252b3806de4cb09f03991d789b83e924a11
Fixes: 119322269
Test: atest KeyguardDisableHandlerTest CtsActivityManagerDeviceTestCases:KeyguardTests