Updates buttons in notification menu row.
The new design for the notification menu row with the new interruption
model enabled is to have nothing on the left side, and to have a single
button for interruption settings on the right which opens the guts (instead
of multiple buttons which take immediate action).  This change implements
that design and removes the mechanism for buttons taking immediate action.

Test: manual
Bug: 116622974
Change-Id: Ibcb2e10453ab490e8de69b00f37aed498a314808