Support wallpaper on secondary displays.(5/N)
1. Pull DisplayData from WallpaperData, so the rendering size will only
related to the display.
2. In the previous CL, we provided a new property "supportsMultipleDisplays"
that lets the WallpaperManagerService or wallpaper selector know that the
component can be rendered on each display at the same time. In this CL we
provide a fallback mechanism. If the selected wallpaper service does not
support multiple displays, we will display the default image wallpaper on
the secondary display.
3. Add private access for some fields and methods.(by IDE reminder)

Bug: 115486823
Test: atest WallpaperManagerTest
Test: atest WmTests
Test: atest ActivityManagerMultiDisplayTests
Change-Id: I2fa6ad5dc39a84e317680399e70178ca258a22ae